requirements.txt
openai==1.30.0
faiss-cpu==1.8.0
fastapi==0.111.0
uvicorn==0.30.0
python-dotenv==1.0.1
numpy==1.26.0
pydantic==2.7.0
tiktoken==0.7.0
tqdm==4.66.0
rank-bm25==0.2.2
orjson==3.10.0